Moved test sources into /src - Finally the IDE will resolve the "paths"-fueled imports.
