+++ src/test/frontend/ArticleEditForm/FormReset.test.js
@@ -9,7 +9,7 @@ const {
 const {
 	log,
 	openPage,
-} = require('../../test-tools-frontend');
+} = require('../test-tools-frontend');
 
 const timeToWait = 5000;
