+++ src/test/frontend/FeedbackFormContainer/Structure.test.js
@@ -1,6 +1,6 @@
 const {
 	openPage
-} = require('../../test-tools-frontend');
+} = require('../test-tools-frontend');
 
 const article_el_1 = 'article#article-el-1';
 const article_el_2 = 'article#article-el-2';
